Se agrego logger manager y se actualizaron los ABM y gestor de producto

+++ Back-end/Api Template/Api Template/Models/BD/inventary.cs
@@ -16,7 +16,7 @@ namespace Api_control_comercio.Models.BD
     {
         public System.Guid inventary_id { get; set; }
         public Nullable<System.Guid> raw_material_id { get; set; }
-        public Nullable<int> quantity { get; set; }
+        public Nullable<decimal> quantity { get; set; }
         public Nullable<System.Guid> physical_location_id { get; set; }
         public Nullable<System.DateTime> creation_date { get; set; }
         public Nullable<System.DateTime> modification_date { get; set; }
