+++ Back-end/Api Template/Api Template/Models/BD/payment_employee.cs
@@ -16,7 +16,7 @@ namespace Api_control_comercio.Models.BD
     {
         public System.Guid payment_employee_id { get; set; }
         public Nullable<System.Guid> employee_id { get; set; }
-        public Nullable<int> payment_employee_price { get; set; }
+        public Nullable<decimal> payment_employee_price { get; set; }
         public Nullable<System.DateTime> creation_date { get; set; }
         public Nullable<System.DateTime> modification_date { get; set; }
     }
