+++ Back-end/Api Template/Api Template/Models/BD/payment_suppliers.cs
@@ -17,8 +17,8 @@ namespace Api_control_comercio.Models.BD
         public System.Guid payment_suppliers_id { get; set; }
         public Nullable<System.Guid> supplier_id { get; set; }
         public Nullable<System.Guid> physical_location_id { get; set; }
-        public Nullable<int> payment_suppliers_cost { get; set; }
-        public Nullable<int> payment_suppliers_pay { get; set; }
+        public Nullable<decimal> payment_suppliers_cost { get; set; }
+        public Nullable<decimal> payment_suppliers_pay { get; set; }
         public Nullable<System.DateTime> creation_date { get; set; }
         public Nullable<System.DateTime> modification_date { get; set; }
     }
