+++ Back-end/Api Template/Api Template/Models/BD/product_rawmaterial.cs
@@ -17,7 +17,7 @@ namespace Api_control_comercio.Models.BD
         public System.Guid product_rawmaterial_id { get; set; }
         public Nullable<System.Guid> product_id { get; set; }
         public Nullable<System.Guid> raw_material_id { get; set; }
-        public Nullable<int> quantity { get; set; }
+        public Nullable<decimal> quantity { get; set; }
         public Nullable<System.DateTime> creation_date { get; set; }
         public Nullable<System.DateTime> modification_date { get; set; }
     }
