+++ Back-end/Api Template/Api Template/Models/BD/sale.cs
@@ -17,7 +17,7 @@ namespace Api_control_comercio.Models.BD
         public System.Guid sale_id { get; set; }
         public Nullable<System.Guid> physical_location_id { get; set; }
         public Nullable<System.Guid> employee_id { get; set; }
-        public Nullable<int> sale_price { get; set; }
+        public Nullable<decimal> sale_price { get; set; }
         public Nullable<System.Guid> payment_method_id { get; set; }
         public Nullable<System.DateTime> creation_date { get; set; }
         public Nullable<System.DateTime> modification_date { get; set; }
