+++ Back-end/Api Template/Api Template/Models/BD/sale_order.cs
@@ -17,8 +17,8 @@ namespace Api_control_comercio.Models.BD
         public System.Guid sale_order_id { get; set; }
         public Nullable<System.Guid> sale_id { get; set; }
         public Nullable<System.Guid> product_id { get; set; }
-        public Nullable<int> quantity { get; set; }
-        public Nullable<int> sale_order_price { get; set; }
+        public Nullable<decimal> quantity { get; set; }
+        public Nullable<decimal> sale_order_price { get; set; }
         public Nullable<System.DateTime> creation_date { get; set; }
         public Nullable<System.DateTime> modification_date { get; set; }
     }
